+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [1.1.5]
 ### Fixes
 - Do not send multiple touch requests when one is pending